+++ bot/states/movie_states.py
@@ -1,6 +1,0 @@
-from aiogram.fsm.state import StatesGroup, State
-
-class MovieStates(StatesGroup):
-    waiting_for_title = State()
-    waiting_for_choice = State()
-    waiting_for_confirmation = State()